TEA-18: Refacotor - Multiple model system - 2
* DebugModel integration
* sonarcube and code review fix

+++ Teapot/src/Models/ModelManager.cpp
@@ -2,9 +2,9 @@
 
 namespace Teapot
 {
-	unsigned int ModelManager::GetModelVectorSize()
+	size_t ModelManager::GetModelVectorSize()
 	{
-		int size = 0;
+		size_t size = 0;
 		for (const auto& modelVec : s_AllModels)
 		{
 			size += modelVec->size();
